feat(mypage): setting view vc로부터 분리 작업 진행
- 일관된 뷰 구성 방법 선택
